+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Add JWT - Part 2 - Service
 + Add JWT - Part 3 - Entity (no related to database) + Response
 + Add JWT - Part 4 - Override websecurityconfig for authenticationManager + external support for configuration
++ Add JWT - Part 5 - LoginService
 
 ## 0.3.0
 + Upgrade dependencies